Rename package for mixin addons
