+++ src/main/java/io/github/srdjanv/tweakedpetroleum/api/crafting/TweakedPumpjackHandler.java
@@ -4,7 +4,7 @@ import blusunrize.immersiveengineering.api.DimensionChunkCoords;
 import flaxbeard.immersivepetroleum.api.crafting.PumpjackHandler;
 import io.github.srdjanv.tweakedlib.api.powertier.PowerTier;
 import io.github.srdjanv.tweakedlib.api.powertier.PowerTierHandler;
-import io.github.srdjanv.tweakedpetroleum.api.ihelpers.IReservoirType;
+import io.github.srdjanv.tweakedpetroleum.api.mixins.IReservoirType;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
