+++ src/main/java/io/github/srdjanv/tweakedpetroleum/api/mixins/IPumpjackAddons.java
@@ -1,4 +1,4 @@
-package io.github.srdjanv.tweakedpetroleum.api.ihelpers;
+package io.github.srdjanv.tweakedpetroleum.api.mixins;
 
 import io.github.srdjanv.tweakedpetroleum.api.crafting.TweakedPumpjackHandler;
